requirements.txt
flask>=3.0,<4
tzdata>=2024.1
python-dotenv>=1.0
gspread>=6.0
google-auth>=2.0
gunicorn>=22.0
pytest>=8.0
